initial commit
After asynchronously loading the columns of interest into memory, a
ListViewAdapter is created with the list of Movies. The text in each
list item is rendered directly whereas the Image downloaded is offloaded
to a rather complete (and over-kill for this assignment) process of
downloading the images and storing them to disk to act as a cache.
Before the image is loaded, a loading spinner is displayed in its place.
This solution will work with far more movie items than was given in the
example as list item view memory is constantly re-used.
I created a simple ScrollView based view in a new Activity to render the
detail view, again using the ImageLoader to fetch the image but this
time the full size one.